+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `sandstone/Slider` prop `keyFrequency` to control the accelerating speed when key hold
+
 ### Fixed
 
 - `sandstone/FormCheckboxItem` to show correct color for `slotBefore` icon in disabled state when focused
@@ -17,7 +21,6 @@ The following is a curated list of changes in the Enact sandstone module, newest
 - `sandstone/Dropdown` number type `width` prop
 - `sandstone/Item` public class names `itemContent`, `content`, and `label`
 - `sandstone/Scroller` prop `scrollbarTrackCss` to customize scroll track and thumb style
-- `sandstone/Slider` prop `keyFrequency` to control the accelerating speed when key hold
 
 ### Fixed